+++ src/vue/modules/create-sale-form/components/full-description-step-form.vue
@@ -38,7 +38,14 @@
         <span>
           {{ 'create-sale-form.full-description-lbl' | globalize }}
         </span>
-        <markdown-field v-model="form.description" />
+        <markdown-field
+          v-model="form.description"
+          @blur="touchField('form.description')"
+          :error-message="getFieldErrorMessage(
+            'form.description',
+            { length: DESCRIPTION_MAX_LENGTH }
+          )"
+        />
       </div>
     </div>
 
@@ -73,6 +80,11 @@ import FormMixin from '@/vue/mixins/form.mixin'
 
 import { CreateSaleRequest } from '../wrappers/create-sale-request'
 
+import {
+  maxLength,
+} from '@validators'
+
+const DESCRIPTION_MAX_LENGTH = 8000
 const EVENTS = {
   submit: 'submit',
   updateIsDisabled: 'update:isDisabled',
@@ -92,6 +104,7 @@ export default {
       youtubeVideo: '',
       description: '',
     },
+    DESCRIPTION_MAX_LENGTH,
   }),
 
   computed: {
@@ -106,6 +119,16 @@ export default {
   created () {
     if (this.request) {
       this.populateForm()
+    }
+  },
+
+  validations () {
+    return {
+      form: {
+        description: {
+          maxLength: maxLength(DESCRIPTION_MAX_LENGTH),
+        },
+      },
     }
   },